Refactor deposit and withdraw form
