+++ src/components/DepositWidget/Form.tsx
@@ -1,0 +1,204 @@
+import React, { useState, useEffect, ChangeEvent, ReactNode } from 'react'
+import styled from 'styled-components'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { IconDefinition } from '@fortawesome/free-solid-svg-icons'
+
+import { TokenBalanceDetails } from 'types'
+import { formatAmountFull, parseAmount } from 'utils'
+import BN from 'bn.js'
+
+const Wrapper = styled.tr`
+  background-color: #f7f0ff;
+
+  td {
+    position: relative;
+  }
+
+  span.symbol {
+    color: #b02ace;
+  }
+
+  h4 {
+    margin: 0;
+    font-size: 1.3em;
+  }
+
+  ul {
+    align-items: center;
+    list-style: none;
+    text-align: left;
+
+    li {
+      display: block;
+      margin: 0 auto;
+      width: 25em;
+    }
+
+    li > label {
+      width: 8em;
+      color: #6c0084;
+      font-weight: bold;
+      text-align: right;
+    }
+
+    p.error {
+      color: red;
+      padding: 0 0 0.5em 10em;
+      margin: 0;
+    }
+
+    div.wallet {
+      display: inline-block;
+      text-align: center;
+      position: relative;
+
+      a.max {
+        display: inline-block;
+        margin-left: 0.5em;
+        position: absolute;
+        top: 1.3em;
+        right: -3em;
+      }
+    }
+
+    li > label {
+      display: inline-block;
+      height: 100%;
+    }
+
+    .buttons {
+      text-align: center;
+      padding-top: 1em;
+      button {
+        width: 7em;
+        margin-left: 1.2em;
+      }
+    }
+  }
+
+  .times {
+    position: absolute;
+    top: 0;
+    right: 0;
+    text-decoration: none;
+    font-size: 2em;
+    display: inline-block;
+    padding: 0 0.5em 0 0;
+  }
+`
+
+export interface FormProps {
+  tokenBalances: TokenBalanceDetails
+  title: ReactNode
+  totalAmountLabel: string
+  totalAmount: BN
+  inputLabel: string
+  submitBtnLabel: string
+  submitBtnIcon: IconDefinition
+  onSubmit: () => Promise<void>
+  onClose: () => void
+}
+
+interface Errors {
+  depositAmount: string
+}
+
+function _validateDeposit(totalAmount: BN, depositAmount: string, decimals: number): string {
+  if (!depositAmount) {
+    return 'Required amount'
+  }
+
+  const amount = parseAmount(depositAmount, decimals)
+  if (!amount) {
+    return 'Invalid amount'
+  }
+
+  if (amount.gt(totalAmount)) {
+    return 'Insuficient balance'
+  }
+
+  return null // no error
+}
+
+export const Form: React.FC<FormProps> = (props: FormProps) => {
+  const { symbol, decimals } = props.tokenBalances
+  const { title, totalAmount, totalAmountLabel, inputLabel, submitBtnLabel, submitBtnIcon } = props
+  const [depositAmount, setDepositAmount] = useState('')
+  const [validatorActive, setValidatorActive] = useState(false)
+  const [errors, setErrors] = useState<Errors>({
+    depositAmount: null,
+  })
+
+  const cancelForm = (): void => {
+    setDepositAmount('')
+    props.onClose()
+  }
+
+  useEffect(() => {
+    if (validatorActive) {
+      // Verify on every deposit change
+      const errorMsg = _validateDeposit(totalAmount, depositAmount, decimals)
+      const newErrors = {
+        ...errors,
+        depositAmount: errorMsg,
+      }
+      setErrors(newErrors)
+      // const hasErrors = Object.keys(newErrors).some(key => !!newErrors[key])
+      // return !hasErrors
+    }
+  }, [depositAmount, validatorActive])
+
+  return (
+    <Wrapper>
+      <td colSpan={6}>
+        <a className="times" onClick={cancelForm}>
+          &times;
+        </a>
+        <h4>{title}</h4>
+        <ul>
+          <li>
+            <label>{totalAmountLabel}</label>
+            <div className="wallet">
+              <input type="text" value={formatAmountFull(totalAmount, decimals)} disabled />
+              <br />
+              <a className="max" onClick={(): void => setDepositAmount(formatAmountFull(totalAmount, decimals, false))}>
+                <small>Use Max</small>
+              </a>
+            </div>
+          </li>
+          <li>
+            <label>{inputLabel}</label>
+            <input
+              type="text"
+              value={depositAmount}
+              onChange={(e: ChangeEvent<HTMLInputElement>): void => setDepositAmount(e.target.value)}
+              placeholder={symbol + ' amount'}
+            />
+            {errors.depositAmount && <p className="error">{errors.depositAmount}</p>}
+          </li>
+          <li className="buttons">
+            <a onClick={cancelForm}>Cancel</a>
+            <button
+              type="button"
+              className="success"
+              disabled={!!errors.depositAmount}
+              onClick={(): void => {
+                setValidatorActive(true)
+                // TODO: Improve. Do not do 2 times the validation
+                const error = _validateDeposit(totalAmount, depositAmount, decimals)
+                if (!error) {
+                  props.onSubmit()
+                }
+              }}
+            >
+              <FontAwesomeIcon icon={submitBtnIcon} />
+              &nbsp; {submitBtnLabel}
+            </button>
+          </li>
+        </ul>
+      </td>
+    </Wrapper>
+  )
+}
+
+export default Form
